Catch more global variables in Export
This is perhaps a little controversial.  At the moment we only export global
functions from GLOBAL_FUNCTION_NAMES, that is, the ones defined using
InstallGlobalFunction.  But lots of functions are defined using just BindGlobal,
meaning that they go missing and end up in MitM_TypesToJson(...)[2].

Lots of these functions are internals that we don't want to tell the world about
anyway, but some are important user functions like DihedralGroup.  This commit
adds all such functions to the export, so long as they don't have underscores or
all caps (characteristic of internal functions).  Perhaps this filter should be
modified later, but this will do for now.

https://gl.mathhub.info/ODK/GAP/tree/master/gap shows the results in action.